move selFile func to separate file to be used in maproom init
