+++ subseas_flex_fcst/maproom.py
@@ -19,6 +19,7 @@ import cptio
 import xarray as xr
 from scipy.stats import t, norm, rankdata
 import pandas as pd
+import predictions
 
 CONFIG = pyaconf.load(os.environ["CONFIG"])
 
@@ -48,30 +49,18 @@ APP.title = "Sub-Seasonal Forecast"
 
 APP.layout = layout.app_layout
 
-def selFile(dataPath, filePattern, leadTime, startDate):
-    #filesNameList = glob.glob(f'{dataPath}/{filePattern}')
-    pattern = f"{startDate}_{leadTime}"
-    fullPath = f"{dataPath}/{filePattern}"
-    print(fullPath)
-    fileName = fullPath.replace("*",pattern)
-    print(fullPath)
-    fileSelected = cptio.open_cptdataset(fileName)
-    startDT = datetime.strptime(startDate, "%b-%d-%Y")
-    fileSelected = fileSelected.expand_dims({"S":[startDT]})
-    return fileSelected
-
 def read_cptdataset(leadTime, startDate, y_transform=CONFIG["y_transform"]): #add leadTime and startDate as inputs
-    fcst_mu = selFile(DATA_PATH, CONFIG["forecast_mu_filePattern"], leadTime, startDate)
+    fcst_mu = predictions.selFile(DATA_PATH, CONFIG["forecast_mu_filePattern"], leadTime, startDate)
     fcst_mu_name = list(fcst_mu.data_vars)[0]
     fcst_mu = fcst_mu[fcst_mu_name]
-    fcst_var = selFile(DATA_PATH, CONFIG["forecast_var_filePattern"], leadTime, startDate)
+    fcst_var = predictions.selFile(DATA_PATH, CONFIG["forecast_var_filePattern"], leadTime, startDate)
     fcst_var_name = list(fcst_var.data_vars)[0]
     fcst_var = fcst_var[fcst_var_name]
-    obs = (selFile(DATA_PATH, CONFIG["obs_filePattern"], leadTime, startDate)).squeeze()
+    obs = (predictions.selFile(DATA_PATH, CONFIG["obs_filePattern"], leadTime, startDate)).squeeze()
     obs_name = list(obs.data_vars)[0]
     obs = obs[obs_name]
     if y_transform:
-        hcst = (selFile(DATA_PATH, CONFIG["hcst_filePattern"], leadTime, startDate)).squeeze()
+        hcst = (predictions.selFile(DATA_PATH, CONFIG["hcst_filePattern"], leadTime, startDate)).squeeze()
         hcst_name = list(hcst.data_vars)[0]
         hcst = hcst[hcst_name]
     else:
